@@ -1,45 +1,47 @@
 {
   "name": "counters",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "scripts": {
-    "build": "ng build",
-    "deploy": "ng build --prod; firebase deploy",
+    "build": "ng build --prod",
+    "deploy": "npm run build; firebase deploy",
     "lint": "ng lint",
     "ng": "ng",
-    "open": "npm run start --open",
+    "open": "ng serve --open=true",
     "start": "ng serve"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/fire": "^5.1.3",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@swimlane/ngx-charts": "^10.1.0",
-    "core-js": "^2.6.5",
-    "firebase": "^5.11.1",
+    "@angular-devkit/build-angular": "^0.803.12",
+    "@angular/animations": "~8.2.11",
+    "@angular/cdk": "8.2.3",
+    "@angular/common": "~8.2.11",
+    "@angular/compiler": "~8.2.11",
+    "@angular/core": "~8.2.11",
+    "@angular/fire": "^5.2.1",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "~8.2.11",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.11",
+    "@angular/platform-browser-dynamic": "~8.2.11",
+    "@angular/pwa": "^0.13.9",
+    "@angular/router": "~8.2.11",
+    "@angular/service-worker": "~8.2.11",
+    "@swimlane/ngx-charts": "^12.0.1",
+    "firebase": "^7.3.0",
+    "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.5.2",
-    "zone.js": "^0.8.29"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.2.4",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/node": "^10.12.30",
-    "codelyzer": "~4.5.0",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.12.1",
-    "typescript": "~3.2.4"
+    "@angular/cli": "~8.3.12",
+    "@angular/compiler-cli": "~8.2.11",
+    "@angular/language-service": "~8.2.11",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~5.0.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
